+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master
 
+## 0.2.0
+
 * warning is not sticky anymore - hanneskaeufler
 * add `fail_for_todos` to fail instead of warn - hanneskaeufler
 * find multiple todos in the same diff - hanneskaeufler